services/tweet-processor/requirements.txt
dapr-ext-grpc==1.6.0